+++ src/main/features/response/routes/check-and-send.ts
@@ -150,11 +150,11 @@ export default express.Router()
         await claimStoreClient.saveResponseForUser(claim, draft, mediationDraft, directionsQuestionnaireDraft, user)
         await draftService.delete(draft.id, user.bearerToken)
 
-        if (mediationDraft.id) {
+        if (draft.document.response.type !== ResponseType.FULL_ADMISSION && mediationDraft.id) {
           await draftService.delete(mediationDraft.id, user.bearerToken)
         }
 
-        if (FeatureToggles.isEnabled('directionsQuestionnaire') && (draft.document.response.type === ResponseType.DEFENCE || draft.document.response.type === ResponseType.PART_ADMISSION)) {
+        if (FeatureToggles.isEnabled('directionsQuestionnaire') && directionsQuestionnaireDraft.id && (draft.document.response.type === ResponseType.DEFENCE || draft.document.response.type === ResponseType.PART_ADMISSION)) {
           await draftService.delete(directionsQuestionnaireDraft.id, user.bearerToken)
         }
